primera modificacion con eliminacion JAIR

+++ resources/views/estado/index.blade.php
@@ -37,20 +37,27 @@
                                               
                                             </tr>
                                     </tfoot>
+                                    @if(count($estados) > 0)
                                         @foreach($estados as $estado)
                                         <tbody>
                                         <td>{{$estado->id_estado}}</td>
                                        <td>{{$estado->nombre}}</td>
                                        <td>
+                                       @if($estado->deleted_at =="")
                                         {!!link_to_route('estado.edit', $title = 'Editar', $parameters = $estado->id_estado, $attributes = ['class'=>'btn btn-success btn-flat btn-addon m-b-10 m-l-5'])!!} 
                                     
                                         {!!Form::open(['route' => ['estado.destroy',$estado->id_estado],'method'=>'DELETE'])!!} 
                                         {!!Form::submit('Eliminar',['class'=>'btn btn-inverse waves-effect waves-light'  ])!!}
                                         {!!Form::close()!!}
+                                        @else
+                                        {!!link_to_route('estado.show', $title = 'Restaurar', $parameters =  $estado->id_estado, $attributes = ['class'=>'btn btn-info btn-flat btn-addon m-b-10 m-l-5'] )!!} 
+                                        
+                                        @endif
 
                                        </td>
                                         </tbody>
                                        @endforeach
+                                       @endif
                                    </table>
                                 </div>
                                </div>
